[refactor] fixture 네임 변경 => fixtures
